Default vg init to all 3 agents, inline install-skills, trim install.md
- Default --agent is now 'claude-code,cursor,codex'. vercel-labs/skills
  stores skills once in ./.agents/skills/ and symlinks into each agent
  dir, so the extra agents are ~free on disk.
- install-skills.ts (20 lines, only used by init) inlined into
  init.ts. Drops redundant consola logging — skills CLI's own stdio
  is already inherited.
- install.md: remove the "older docs to ignore" paragraph and the
  "Notes" section; drop per-agent usage variants since the default
  covers them.
