+++ apps/cli/src/commands/init.ts
@@ -1,9 +1,25 @@
+import { spawn } from "node:child_process";
+
 import { defineCommand } from "citty";
 
-import { installSkills } from "../lib/install-skills.js";
+const REPO = "kyh/vibedgames";
+const DEFAULT_AGENTS = "claude-code,cursor,codex";
+const description = "Install vibedgames skills into your project";
 
-const description =
-  "Install vibedgames skills into your project (delegates to `npx skills add kyh/vibedgames` from vercel-labs/skills)";
+const runSkillsAdd = async (agents: string[], global: boolean, yes: boolean) => {
+  const args = ["-y", "skills", "add", REPO];
+  for (const agent of agents) args.push("-a", agent);
+  if (global) args.push("-g");
+  if (yes) args.push("-y");
+
+  const npx = process.platform === "win32" ? "npx.cmd" : "npx";
+  const code = await new Promise<number>((resolve, reject) => {
+    const child = spawn(npx, args, { stdio: "inherit" });
+    child.on("exit", (c) => resolve(c ?? 1));
+    child.on("error", reject);
+  });
+  if (code !== 0) throw new Error(`skills exited with code ${code}`);
+};
 
 export const initCommand = defineCommand({
   meta: { name: "init", description },
@@ -11,9 +27,9 @@ export const initCommand = defineCommand({
     agent: {
       type: "string",
       description:
-        "Target agent (claude-code, cursor, codex, windsurf, ...). Pass '*' to pick all.",
+        "Comma-separated target agents. Default installs for Claude Code, Cursor, and Codex (symlinked from a shared .agents/skills/ dir). Pass '*' for every supported agent.",
       alias: "a",
-      default: "claude-code",
+      default: DEFAULT_AGENTS,
     },
     global: {
       type: "boolean",
@@ -28,12 +44,13 @@ export const initCommand = defineCommand({
       alias: "y",
     },
   },
-  run: ({ args }) =>
-    installSkills({
-      agent: args.agent,
-      global: args.global,
-      yes: args.yes,
-    }),
+  run: ({ args }) => {
+    const agents = args.agent
+      .split(",")
+      .map((s) => s.trim())
+      .filter(Boolean);
+    return runSkillsAdd(agents, args.global, args.yes);
+  },
 });
 
 export const skillsInstallCommand = defineCommand({
