+++ apps/web/src/lib/install.md
@@ -10,20 +10,15 @@ built-in multiplayer. Execute the steps below.
 npx vibedgames init
 ```
 
-That's it. By default this installs skills for Claude Code
-non-interactively. If you're a different agent, pass `--agent`:
+Installs skills for Claude Code, Cursor, and Codex in one shot. Under
+the hood this wraps `npx skills add kyh/vibedgames` from
+[vercel-labs/skills](https://github.com/vercel-labs/skills): the skills
+live once in `./.agents/skills/` and are symlinked into `.claude/skills/`,
+`.cursor/skills/`, and `.codex/skills/`. Windows without symlink support
+falls back to copies.
 
-- Cursor → `npx vibedgames init --agent cursor`
-- Codex (OpenAI) → `npx vibedgames init --agent codex`
-- Anything else → `npx vibedgames init --agent <name>`
-  (supports 45+ agents — see
-  [vercel-labs/skills](https://github.com/vercel-labs/skills))
-
-Under the hood, `vibedgames init` delegates to `npx skills add
-kyh/vibedgames`: copies the skills to `./.agents/skills/` and symlinks
-them into your agent's skills dir (`.claude/skills/`, `.cursor/skills/`,
-`.codex/skills/`, etc). Windows without symlink support falls back to
-copies automatically.
+If you're a different agent, pass `--agent <name>` (supports 45+ agents
+— see vercel-labs/skills).
 
 ## 2. Authenticate
 
@@ -31,9 +26,8 @@ copies automatically.
 npx vibedgames login
 ```
 
-Device-code flow: the CLI prints an 8-character code and a URL. Ask the
-user to open the URL in their browser and confirm the code. The CLI
-polls and stores the token once confirmed.
+Device-code flow: the CLI prints an 8-char code and a URL. Ask the user
+to open the URL in their browser and confirm the code.
 
 Verify with `npx vibedgames whoami`.
 
@@ -50,10 +44,3 @@ Next commands they can run:
 - `npx skills update` — refresh the vibedgames skills later
 
 Docs: https://vibedgames.com
-
-## Notes
-
-- `vibedgames` is our CLI on npm (binary `vg`). `npx vibedgames init`
-  and `vg init` are the same command.
-- Older docs may mention Claude-Code-specific `/plugin` commands —
-  ignore those, they're superseded by the flow above.
